submission to asset_group dir rename
