+++ app/modules/asset_groups/models.py
